Fixed major bug and finished project.
After multiple hours of troubleshooting I used ChatGPT to help me troubleshoot the issue and found a solution for the "people1" line bug. Added the rest of the features of the API and added some comments. Gonna continue adding comments and maybe expand the utility of the API with more database utility features.